+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
+## [0.0.3]
+
+### Changed
+
+- Refactor config
+
+### Added
+
+- Option to load / save config (private key)
+- Balance check
+
 ## [0.0.2]
 
 ### Fixed